+++ frontend/monaco-ide-starter/package.json
@@ -7,9 +7,10 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@monaco-editor/react": "^4.4.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "@monaco-editor/react": "^4.4.6"
+    "react-router-dom": "^7.7.0"
   },
   "devDependencies": {
     "vite": "^5.0.0"
