+++ frontend/monaco-ide-starter/src/App.jsx
@@ -1,28 +1,16 @@
 import React from "react";
-
-import { useState } from "react";
-import CodeEditor from "./components/CodeEditor";
-import OutputBox from "./components/OutputBox";
-import LanguageSelector from "./components/LanguageSelector";
-import RunButton from "./components/RunButton";
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import Home from "./components/Home";
+import IDE from "./components/IDE";
 
 function App() {
-  const [code, setCode] = useState("// Start coding...");
-  const [output, setOutput] = useState("");
-  const [language, setLanguage] = useState("javascript");
-
-  const handleRun = () => {
-    setOutput("Running... (connect backend here)");
-  };
-
   return (
-    <div style={{ padding: "1rem" }}>
-      <h1>Online IDE</h1>
-      <LanguageSelector language={language} setLanguage={setLanguage} />
-      <CodeEditor code={code} setCode={setCode} language={language} />
-      <RunButton onClick={handleRun} />
-      <OutputBox output={output} />
-    </div>
+    <Router>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/ide" element={<IDE />} />
+      </Routes>
+    </Router>
   );
 }
